refactor: create library for storing identity and enrolment interfaces used by frontend and backend
