+++ apps/dsb-client-gateway-api/src/app/modules/identity/dto/claims-response.dto.ts
@@ -1,4 +1,4 @@
-import { RoleState } from '../../storage/storage.interface';
+import { RoleState } from '@dsb-client-gateway/dsb-client-gateway/identity/models';
 
 export class ClaimResponseDto {
   namespace: string;
