+++ apps/dsb-client-gateway-api/src/app/modules/identity/service/identity.service.ts
@@ -1,17 +1,18 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { EthersService } from '../../utils/service/ethers.service';
-import {
-  Enrolment,
-  Identity,
-  RoleState,
-} from '../../storage/storage.interface';
 import { IamService } from '../../iam-service/service/iam.service';
 import { SecretsEngineService } from '../../secrets-engine/secrets-engine.interface';
 import { NoPrivateKeyException } from '../../storage/exceptions/no-private-key.exception';
 import { EnrolmentRepository } from '../../storage/repository/enrolment.repository';
 import { IdentityRepository } from '../../storage/repository/identity.repository';
 import { EnrolmentService } from '../../enrolment/service/enrolment.service';
-import { Claims, IdentityWithEnrolment } from '../identity.interface';
+import {
+  Claims,
+  Enrolment,
+  Identity,
+  IdentityWithEnrolment,
+  RoleState
+} from '@dsb-client-gateway/dsb-client-gateway/identity/models';
 import { Claim } from 'iam-client-lib';
 
 @Injectable()
