+++ apps/dsb-client-gateway-api/src/app/modules/storage/repository/identity.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { AbstractLokiRepository } from './abstract-loki.repository';
-import { Identity } from '../storage.interface';
+import { Identity } from '@dsb-client-gateway/dsb-client-gateway/identity/models';
 import { LokiService } from '../service/loki.service';
 
 @Injectable()
