+++ libs/dsb-client-gateway/identity/models/README.md
@@ -1,0 +1,7 @@
+# dsb-client-gateway-identity-models
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test dsb-client-gateway-identity-models` to execute the unit tests via [Jest](https://jestjs.io).
