+++ libs/dsb-client-gateway/identity/models/jest.config.js
@@ -1,0 +1,16 @@
+module.exports = {
+  displayName: 'dsb-client-gateway-identity-models',
+  preset: '../../../../jest.preset.js',
+  globals: {
+    'ts-jest': {
+      tsconfig: '<rootDir>/tsconfig.spec.json',
+    },
+  },
+  testEnvironment: 'node',
+  transform: {
+    '^.+\\.[tj]sx?$': 'ts-jest',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  coverageDirectory:
+    '../../../../coverage/libs/dsb-client-gateway/identity/models',
+};
