+++ libs/dsb-client-gateway/identity/models/project.json
@@ -1,0 +1,23 @@
+{
+  "root": "libs/dsb-client-gateway/identity/models",
+  "sourceRoot": "libs/dsb-client-gateway/identity/models/src",
+  "projectType": "library",
+  "targets": {
+    "lint": {
+      "executor": "@nrwl/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["libs/dsb-client-gateway/identity/models/**/*.ts"]
+      }
+    },
+    "test": {
+      "executor": "@nrwl/jest:jest",
+      "outputs": ["coverage/libs/dsb-client-gateway/identity/models"],
+      "options": {
+        "jestConfig": "libs/dsb-client-gateway/identity/models/jest.config.js",
+        "passWithNoTests": true
+      }
+    }
+  },
+  "tags": []
+}
