+++ libs/dsb-client-gateway/identity/models/src/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/dsb-client-gateway-identity-models';
+export * from './lib/balance.enum';
+export * from './lib/identity.interface';
