+++ tsconfig.base.json
@@ -18,6 +18,9 @@
     "paths": {
       "@dsb-client-gateway/dsb-client-gateway-errors": [
         "libs/dsb-client-gateway-errors/src/index.ts"
+      ],
+      "@dsb-client-gateway/dsb-client-gateway/identity/models": [
+        "libs/dsb-client-gateway/identity/models/src/index.ts"
       ]
     }
   },
