+++ workspace.json
@@ -5,6 +5,7 @@
     "dsb-client-gateway-api-e2e": "apps/dsb-client-gateway-api-e2e",
     "dsb-client-gateway-errors": "libs/dsb-client-gateway-errors",
     "dsb-client-gateway-frontend": "apps/dsb-client-gateway-frontend",
-    "dsb-client-gateway-frontend-e2e": "apps/dsb-client-gateway-frontend-e2e"
+    "dsb-client-gateway-frontend-e2e": "apps/dsb-client-gateway-frontend-e2e",
+    "dsb-client-gateway-identity-models": "libs/dsb-client-gateway/identity/models"
   }
 }
